- refactor to expose fee estimation - send beacon confirmations and errors
