+++ src/reducers/app/thunks.ts
@@ -2,10 +2,12 @@ import {
     ConseilQueryBuilder,
     ConseilOperator,
     ConseilSortDirection,
+    Delegation,
     TezosConseilClient,
-    TezosNodeReader,
-    TezosNodeWriter,
+    TezosConstants,
     TezosMessageUtils,
+    TezosNodeReader,
+    TezosNodeWriter
 } from 'conseiljs';
 import {Dispatch} from 'redux';
 
@@ -189,10 +191,7 @@ export const sendTransaction = () => async (
     }
 };
 
-export const sendDelegation = () => async (
-    dispatch: any,
-    getState: () => State,
-) => {
+export const sendDelegation = () => async (dispatch: any, getState: () => State) => {
     try {
         const tezosUrl = config[0].nodeUrl; // TODO: getState().config
         const address = getState().app.delegateAddress; // TODO do not use state, use parameters
@@ -351,3 +350,108 @@ export const getBakerDetails = async (address: string): Promise<BakerInfo> => { 
     }
     return {address, name: '', fee: 0, logoUrl: '', estimatedRoi: 0};
 };
+
+export const estimateOperationFee = async (operations: any, secretKey: any): Promise<number> => {
+    try {
+        const keyStore = await KeyStoreUtils.restoreIdentityFromSecretKey(secretKey);
+        const tezosUrl = config[0].nodeUrl;
+
+        const formedOperations: any = await createOperationGroup(operations, tezosUrl, keyStore.publicKeyHash, keyStore.publicKey);
+
+        for (let o of formedOperations) { // TODO: FIX fees and gas for mint operations
+            if (o.parameters.entrypoint === 'mint_OBJKT') {
+                formedOperations[0].gas_limit = '99999';
+                formedOperations[0].fee = `1${formedOperations[0].fee}`;
+            }
+        }
+
+        const estimate = await TezosNodeWriter.estimateOperationGroup(tezosUrl, 'main', formedOperations);
+
+        return Number(estimate.estimatedFee);
+    } catch {
+        return 0; // TODO: present error to user
+    }
+}
+
+export async function createOperationGroup(
+    operations: any[],
+    tezosUrl: string,
+    publicKeyHash: string,
+    publicKey: string) {
+    const networkCounter = await TezosNodeReader.getCounterForAccount(
+        tezosUrl,
+        publicKeyHash,
+    );
+    const formedOperations: any[] = [];
+
+    let counter = networkCounter;
+    for (const o of operations) {
+        counter += 1;
+
+        switch (o.kind) {
+            case 'transaction': {
+                let entrypoint: string | undefined;
+                let parameters: string | undefined;
+
+                try {
+                    entrypoint = o.parameters.entrypoint;
+                    parameters = JSON.stringify(o.parameters.value);
+                } catch {
+                    //
+                }
+
+                try {
+                    if (entrypoint === 'mint_OBJKT') { // TODO: FIX mint operations bytes convert
+                        if (o.parameters.value.args[1].args[0].bytes.length) {
+                            o.parameters.value.args[1].args[0].bytes = Buffer.from(
+                                o.parameters.value.args[1].args[0].bytes,
+                            ).toString('hex');
+                        }
+                        o.parameters.value.args[1].args[0].bytes = o.parameters.value.args[1].args[0].bytes;
+                        parameters = JSON.stringify(o.parameters.value);
+                    }
+                } catch {
+                    //
+                }
+
+                const op = TezosNodeWriter.constructContractInvocationOperation(
+                    publicKeyHash,
+                    counter,
+                    o.destination,
+                    o.amount,
+                    0,
+                    TezosConstants.OperationStorageCap,
+                    TezosConstants.OperationGasCap,
+                    entrypoint,
+                    parameters,
+                );
+
+                formedOperations.push(op);
+
+                break;
+            }
+            case 'delegation': {
+                const op: Delegation = {
+                    kind: 'delegation',
+                    source: publicKeyHash,
+                    fee: '0',
+                    counter: counter.toString(),
+                    storage_limit:
+                        TezosConstants.DefaultDelegationStorageLimit.toString(),
+                    gas_limit:
+                        TezosConstants.DefaultDelegationGasLimit.toString(),
+                    delegate: o.delegate,
+                };
+
+                formedOperations.push(op);
+
+                break;
+            }
+            default: {
+                break;
+            }
+        }
+    }
+
+    return await TezosNodeWriter.appendRevealOperation(tezosUrl, publicKey, publicKeyHash, networkCounter, formedOperations);
+}
